all components organized
