+++ client/components/SingleReview/CommentForm.jsx
@@ -1,7 +1,6 @@
 
 import Button from '@mui/material/Button';
 import Card from '@mui/material/Card';
-import Typography from '@mui/material/Typography';
 import TextField from '@mui/material/TextField';
 import Stack from '@mui/material/Stack';
 import AddCommentIcon from '@mui/icons-material/AddComment';
